requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
pyswip>=0.2.10
google-generativeai>=0.3.2
python-dotenv>=1.0.0
gunicorn>=21.2.0
supabase>=2.3.0
requests>=2.31.0
